chore: refactor the tool name
