+++ packages/mcp-provider-mobile-web/src/provider.ts
@@ -1,5 +1,5 @@
 import { McpProvider, McpTool, Services } from "@salesforce/mcp-provider-api";
-import { NativeCapabilityTool } from "./tools/native-capabilities/nativeCapabilityTool.js";
+import { NativeCapabilityTool } from "./tools/native-capabilities/sf-mobile-native-capability.js";
 import { OfflineAnalysisTool } from "./tools/offline-analysis/sf-mobile-web-offline-analysis.js";
 import { OfflineGuidanceTool } from "./tools/offline-guidance/sf-mobile-web-offline-guidance.js";
 import { nativeCapabilityConfigs } from "./tools/native-capabilities/nativeCapabilityConfig.js";
