+++ packages/mcp-provider-mobile-web/src/tools/native-capabilities/sf-mobile-native-capability.ts
