+++ packages/mcp-provider-mobile-web/test/tools/nativeCapabilityTool.test.ts
@@ -1,5 +1,5 @@
 import { McpToolConfig, ReleaseState, Toolset } from "@salesforce/mcp-provider-api";
-import { NativeCapabilityTool } from "../../src/tools/native-capabilities/nativeCapabilityTool.js";
+import { NativeCapabilityTool } from "../../src/tools/native-capabilities/sf-mobile-native-capability.js";
 import { AppReviewConfig } from "../../src/tools/native-capabilities/nativeCapabilityConfig.js";
 import { SpyTelemetryService } from "../test-doubles.js";
 import { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
